move _extended_xr_methods to base.py
